27-repository-wishrepository-wishlistrepository-profilerepository

+++ src/main/java/com/johanoliverlarsen/wishlist/model/Wish.java
@@ -1,15 +1,16 @@
 package com.johanoliverlarsen.wishlist.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 public class Wish {
-    private int wishId;
+    private Integer wishId;
     private String title;
     private String description;
     private String location;
     private LocalDate date;
-    private double price;
+    private BigDecimal price;
     private String url;
     private List<String> tag;
 
@@ -17,8 +18,8 @@ public class Wish {
         //tom konstruktør
     }
 
-    public Wish(int wishId, String title, String description, String location,
-                LocalDate date, double price, String url, List<String> tag) {
+    public Wish(Integer wishId, String title, String description, String location,
+                LocalDate date, BigDecimal price, String url, List<String> tag) {
         this.wishId = wishId;
         this.title = title;
         this.description = description;
@@ -29,21 +30,21 @@ public class Wish {
         this.tag = tag;
     }
 
-    public int getWishId() { return wishId; }
+    public Integer getWishId() { return wishId; }
     public String getTitle() { return title; }
     public String getDescription() { return description; }
     public String getLocation() { return location; }
     public LocalDate getDate() { return date; }
-    public double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
     public String getUrl() { return url; }
     public List<String> getTag() { return tag; }
 
-    public void setWishId(int wishId) { this.wishId = wishId; }
+    public void setWishId(Integer wishId) { this.wishId = wishId; }
     public void setTitle(String title) { this.title = title; }
     public void setDescription(String description) { this.description = description; }
     public void setLocation(String location) { this.location = location; }
     public void setDate(LocalDate date) { this.date = date; }
-    public void setPrice(double price) { this.price = price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
     public void setUrl(String url) { this.url = url; }
     public void setTag(List<String> tag) { this.tag = tag; }
